# target/classes/static/swagger.yaml
openapi: 3.0.1
info:
  title: MS-PRODUCTS API
  description: Microservice for managing the product catalog - Lab 2
  version: 1.1.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Products
    description: Operations related to product management

paths:
  /api/products:
    get:
      tags: [Products]
      summary: Get all products
      responses:
        '200':
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDTO'
    post:
      tags: [Products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Invalid input or duplicate code

  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product found
        '404':
          description: Product not found

    put:
      tags: [Products]
      summary: Update an existing product
      description: Requires Admin privileges (validated via userId)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: true
          description: ID of the user performing the update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
      responses:
        '200':
          description: Product updated successfully
        '401':
          description: Unauthorized - User is not an admin
        '404':
          description: Product not found

    delete:
      tags: [Products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found

  /api/products/low-stock:
    get:
      tags: [Products]
      summary: Get products with stock below threshold
      parameters:
        - name: threshold
          in: query
          required: true
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Low stock products retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDTO'

components:
  schemas:
    ProductRequestDTO:
      type: object
      required: [name, code, price, stock]
      properties:
        name:
          type: string
          example: "Gaming Laptop"
        code:
          type: string
          pattern: "^PROD-\\d{4}$"
          example: "PROD-1234"
        price:
          type: number
          format: float
          example: 1200.50
        stock:
          type: integer
          example: 50
        description:
          type: string
          example: "High performance laptop"

    ProductResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        price:
          type: number
        stock:
          type: integer
        active:
          type: boolean